Add batch proclamations (32-485 with missing docs)
This commits adds a very big batch of proclamations from 32 all the way
to 485 but has not been checked for each proclamation's existence in
that range. For example, some 480s range docs are missing b/c they were
not on the fsc.gov.et site.

I used the following script to grab the files:
https://gist.github.com/Maktm/c2e46285bb4012f980d8567698f0446b